Include find-conflicts to engine-step
